move retry tests into their own module
